# docs/swagger.yaml
basePath: /api
definitions:
  handlers.AddPlayerToRoomRequest:
    properties:
      username:
        example: john_doe
        maxLength: 25
        type: string
    required:
    - username
    type: object
  handlers.AddPlayerToRoomResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
      room:
        $ref: '#/definitions/handlers.Room'
    type: object
  handlers.CreateRoomRequest:
    properties:
      name:
        example: Room One
        maxLength: 100
        type: string
    required:
    - name
    type: object
  handlers.CreateRoomResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
      room:
        $ref: '#/definitions/handlers.Room'
    type: object
  handlers.Room:
    properties:
      id:
        type: string
      name:
        type: string
      players:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
...
  models.Error:
    properties:
      errorCode:
        type: integer
      errorDescription:
        type: string
    type: object
...
  models.PlayersInRoomResponse:
    properties:
      error:
        $ref: '#/definitions/models.Error'
      players:
        items:
          $ref: '#/definitions/models.User'
        type: array
      room_id:
        example: room123
        type: string
    type: object
...
  models.User:
    properties:
      email:
        type: string
      id:
        type: string
      image:
        type: string
      lastActivityDate:
        type: string
      registrationDate:
        type: string
      team:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API сервис пользователей для взаимодействия с сервисом авторизации
    и сервисом игровых комнат
  title: User Service API
  version: "1.0"
paths:
  /rooms/{room_id}/players:
    get:
      consumes:
      - application/json
      description: Возвращает список игроков в указанной игровой комнате
      parameters:
      - description: Идентификатор комнаты
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PlayersInRoomResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.PlayersInRoomResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.PlayersInRoomResponse'
      summary: Получение информации об игроках в комнате
      tags:
      - room
    post:
      consumes:
      - application/json
      description: Добавляет пользователя в указанную игровую комнату
      parameters:
      - description: Идентификатор комнаты
        in: path
        name: room_id
        required: true
        type: string
      - description: Информация об игроке
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/handlers.AddPlayerToRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.AddPlayerToRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.AddPlayerToRoomResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.AddPlayerToRoomResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.AddPlayerToRoomResponse'
      summary: Добавление игрока в комнату
      tags:
      - room
  /rooms/create:
    post:
      consumes:
      - application/json
      description: Создает новую игровую комнату с заданным именем
      parameters:
      - description: Информация о комнате
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CreateRoomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.CreateRoomResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.CreateRoomResponse'
      summary: Создание новой игровой комнаты
      tags:
      - room
...
swagger: "2.0"